Move more eval code mathics.builtin to mathics.eval
